+++ src/components/index.js
@@ -5,6 +5,8 @@ export { default as HandlerManager } from './HandlerManager';
 export { default as AppCtxMenuProvider } from './MainNav/CurrentApp/AppCtxMenuProvider';
 export { LastVisitedContext, withLastVisited } from './LastVisited';
 export { default as Login } from './Login';
+export { default as Logout } from './Logout';
+export { default as LogoutTimeout } from './LogoutTimeout';
 export { default as MainContainer } from './MainContainer';
 export { default as MainNav } from './MainNav';
 export { default as ModuleContainer } from './ModuleContainer';
@@ -28,5 +30,7 @@ export { default as CheckEmailStatusPage } from './CheckEmailStatusPage';
 export { default as BadRequestScreen } from './BadRequestScreen';
 export { default as NoPermissionScreen } from './NoPermissionScreen';
 export { default as ResetPasswordNotAvailableScreen } from './ResetPasswordNotAvailableScreen';
+export { default as SessionEventContainer } from './SessionEventContainer';
+
 export * from './ModuleHierarchy';
 export * from './Namespace';